Test display activation right after adding it (#2308)
[stage-1]
